Split home into components for each section
