+++ web/src/components/routeHome/Recents.tsx
@@ -1,0 +1,12 @@
+import PageSection from "../PageSection";
+
+const Recents = () => {
+  return (
+    <PageSection
+      className="col-span-3"
+      titleBar="Jump back in..."
+    ></PageSection>
+  );
+};
+
+export default Recents;
